+++ demo/src/app/guides/guides.component.ts
@@ -23,6 +23,8 @@ export class GuidesComponent implements OnInit, OnDestroy {
     validation: require('!!raw-loader!!highlight-loader!markdown-loader!docs/validation.md'),
     'expression-properties': require('!!raw-loader!!highlight-loader!markdown-loader!docs/expression-properties.md'),
     'formly-field-presets': require('!!raw-loader!!highlight-loader!markdown-loader!docs/formly-field-presets.md'),
+    'json-schema': require('!!raw-loader!!highlight-loader!markdown-loader!docs/json-schema.md'),
+    faq: require('!!raw-loader!!highlight-loader!markdown-loader!docs/faq.md'),
     migration: require('!!raw-loader!!highlight-loader!markdown-loader!UPGRADE-6.0.md'),
   };