+++ src/core/src/lib/extensions/field-expression/field-expression.ts
@@ -48,7 +48,7 @@ export class FieldExpressionExtension implements FormlyExtension {
           value$: (expr as Observable<any>).pipe(
             tap((v) => {
               this.evalExpr(field, key, v);
-              field.options.detectChanges(field);
+              field.options._detectChanges(field);
             }),
           ),
         };
@@ -89,9 +89,6 @@ export class FieldExpressionExtension implements FormlyExtension {
         options._hiddenFieldsForCheck = [];
         if (fieldChanged) {
           this.checkExpressions(field);
-          if (field.options && field.options.detectChanges) {
-            field.options.detectChanges(field);
-          }
         }
         checkLocked = false;
       };
@@ -188,7 +185,7 @@ export class FieldExpressionExtension implements FormlyExtension {
   private changeHideState(field: FormlyFieldConfigCache, hide: boolean, resetOnHide: boolean) {
     if (field.fieldGroup) {
       field.fieldGroup
-        .filter((f: FormlyFieldConfigCache) => !f._expressions.hide)
+        .filter((f: FormlyFieldConfigCache) => f && !f._expressions.hide)
         .forEach((f) => this.changeHideState(f, hide, resetOnHide));
     }
 
@@ -226,30 +223,35 @@ export class FieldExpressionExtension implements FormlyExtension {
   }
 
   private evalExpr(field: FormlyFieldConfigCache, prop: string, value: any) {
-    try {
-      let target: any = field;
-      const paths = this._evalExpressionPath(field, prop);
-      const lastIndex = paths.length - 1;
-      for (let i = 0; i < lastIndex; i++) {
-        target = target[paths[i]];
-      }
-
-      target[paths[lastIndex]] = value;
-    } catch (error: any) {
-      error.message = `[Formly Error] [Expression "${prop}"] ${error.message}`;
-      throw error;
-    }
-
-    if (['templateOptions.disabled', 'props.disabled'].includes(prop) && hasKey(field)) {
-      this.changeDisabledState(field, value);
-    }
-
     if (prop.indexOf('model.') === 0) {
       const key = prop.replace(/^model\./, ''),
-        control = field?.key === key ? field.formControl : field.form.get(key);
+        parent = field.fieldGroup ? field : field.parent;
 
+      let control = field?.key === key ? field.formControl : field.form.get(key);
+      if (!control && field.get(key)) {
+        control = field.get(key).formControl;
+      }
+      assignFieldValue({ key, parent, model: field.model }, value);
       if (control && !(isNil(control.value) && isNil(value)) && control.value !== value) {
         control.patchValue(value);
+      }
+    } else {
+      try {
+        let target: any = field;
+        const paths = this._evalExpressionPath(field, prop);
+        const lastIndex = paths.length - 1;
+        for (let i = 0; i < lastIndex; i++) {
+          target = target[paths[i]];
+        }
+
+        target[paths[lastIndex]] = value;
+      } catch (error: any) {
+        error.message = `[Formly Error] [Expression "${prop}"] ${error.message}`;
+        throw error;
+      }
+
+      if (['templateOptions.disabled', 'props.disabled'].includes(prop) && hasKey(field)) {
+        this.changeDisabledState(field, value);
       }
     }