+++ src/core/src/lib/utils.ts
@@ -2,7 +2,7 @@ import { FormlyFieldConfig } from './models';
 import { isObservable } from 'rxjs';
 import { AbstractControl } from '@angular/forms';
 import { FormlyFieldConfigCache } from './models';
-import { ChangeDetectorRef, ComponentRef, TemplateRef, Type } from '@angular/core';
+import { ChangeDetectorRef, ComponentRef, NgZone, TemplateRef, Type, ɵNoopNgZone } from '@angular/core';
 
 export function disableTreeValidityCall(form: any, callback: Function) {
   const _updateTreeValidity = form._updateTreeValidity.bind(form);
@@ -28,7 +28,7 @@ export function getFieldId(formId: string, field: FormlyFieldConfig, index: stri
 }
 
 export function hasKey(field: FormlyFieldConfig) {
-  return !isNil(field.key) && field.key !== '';
+  return !isNil(field.key) && field.key !== '' && (!Array.isArray(field.key) || field.key.length > 0);
 }
 
 export function getKeyPath(field: FormlyFieldConfigCache): string[] {
@@ -169,6 +169,18 @@ export function clone(value: any): any {
     return new Map(value);
   }
 
+  if (value instanceof Uint8Array) {
+    return new Uint8Array(value);
+  }
+
+  if (value instanceof Uint16Array) {
+    return new Uint16Array(value);
+  }
+
+  if (value instanceof Uint32Array) {
+    return new Uint32Array(value);
+  }
+
   // https://github.com/moment/moment/blob/master/moment.js#L252
   if (value._isAMomentObject && isFunction(value.clone)) {
     return value.clone();
@@ -212,7 +224,7 @@ export function defineHiddenProp(field: any, prop: string, defaultValue: any) {
 
 type IObserveFn<T> = (change: { currentValue: T; previousValue?: T; firstChange: boolean }) => void;
 export interface IObserver<T> {
-  setValue: (value: T) => void;
+  setValue: (value: T, emitEvent?: boolean) => void;
   unsubscribe: Function;
 }
 interface IObserveTarget<T> {
@@ -225,7 +237,7 @@ interface IObserveTarget<T> {
   };
 }
 
-export function observeDeep(source: any, paths: string[], setFn: () => void): () => void {
+export function observeDeep<T = any>(source: IObserveTarget<T>, paths: string[], setFn: () => void): () => void {
   let observers: Function[] = [];
 
   const unsubscribe = () => {
@@ -249,7 +261,7 @@ export function observeDeep(source: any, paths: string[], setFn: () => void): ()
   };
 }
 
-export function observe<T = any>(o: IObserveTarget<T>, paths: string[], setFn: IObserveFn<T>): IObserver<T> {
+export function observe<T = any>(o: IObserveTarget<T>, paths: string[], setFn?: IObserveFn<T>): IObserver<T> {
   if (!o._observers) {
     defineHiddenProp(o, '_observers', {});
   }
@@ -273,10 +285,10 @@ export function observe<T = any>(o: IObserveTarget<T>, paths: string[], setFn: I
     state.value = target[key];
   }
 
-  if (state.onChange.indexOf(setFn) === -1) {
+  if (setFn && state.onChange.indexOf(setFn) === -1) {
     state.onChange.push(setFn);
     setFn({ currentValue: state.value, firstChange: true });
-    if (state.onChange.length >= 1) {
+    if (state.onChange.length >= 1 && isObject(target)) {
       const { enumerable } = Object.getOwnPropertyDescriptor(target, key) || { enumerable: true };
       Object.defineProperty(target, key, {
         enumerable,
@@ -294,8 +306,18 @@ export function observe<T = any>(o: IObserveTarget<T>, paths: string[], setFn: I
   }
 
   return {
-    setValue(value: T) {
-      state.value = value;
+    setValue(currentValue: T, emitEvent = true) {
+      if (currentValue === state.value) {
+        return;
+      }
+
+      const previousValue = state.value;
+      state.value = currentValue;
+      state.onChange.forEach((changeFn) => {
+        if (changeFn !== setFn && emitEvent) {
+          changeFn({ previousValue, currentValue, firstChange: false });
+        }
+      });
     },
     unsubscribe() {
       state.onChange = state.onChange.filter((changeFn) => changeFn !== setFn);
@@ -340,4 +362,22 @@ export function markFieldForCheck(field: FormlyFieldConfigCache) {
       ref.markForCheck();
     }
   });
+}
+
+export function isNoopNgZone(ngZone: NgZone) {
+  return ngZone instanceof ɵNoopNgZone;
+}
+
+export function isHiddenField(field: FormlyFieldConfig) {
+  const isHidden = (f: FormlyFieldConfig) => f.hide || f.expressions?.hide || f.hideExpression;
+  let setDefaultValue = !field.resetOnHide || !isHidden(field);
+  if (!isHidden(field) && field.resetOnHide) {
+    let parent = field.parent;
+    while (parent && !isHidden(parent)) {
+      parent = parent.parent;
+    }
+    setDefaultValue = !parent || !isHidden(parent);
+  }
+
+  return !setDefaultValue;
 }