+++ src/ui/bootstrap/textarea/src/textarea.type.ts
@@ -22,8 +22,9 @@ export interface FormlyTextAreaFieldConfig extends FormlyFieldConfig<TextAreaPro
         class="form-control"
         [class.is-invalid]="showError"
         [formlyAttributes]="field"
-      >
-      </textarea>
+        [attr.aria-describedby]="id + '-formly-validation-error'"
+        [attr.aria-invalid]="showError"
+      ></textarea>
     </ng-template>
   `,
   changeDetection: ChangeDetectionStrategy.OnPush,